feat: show enum values as argument in help text (#276)
When options or arguments are parsed into an enum, the help text will now include the values contained in the enum.

Fixes #92

Also made the default constructor public. This allows the DefaultHelpTextGenerator to be instantiated on a per-use basis in order to keep unit tests a isolated from each other as possible.
Without that change, all instances would be the same singleton instance - i.e. a test can affect another test by fiddling with the overridden width.